chm_requirements.txt
# CHM Complete Requirements - All dependencies for production deployment
# This file contains all requirements as specified in CLAUDE.md documentation
# Last Updated: December 2024
# =====================================================================

# Core Framework & Web Server
# ---------------------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Configuration & Environment
# ---------------------------
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database Layer (PostgreSQL + Async)
# ------------------------------------
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
# -------------------------
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7
PyJWT==2.8.0

# HTTP & Networking
# -----------------
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# SNMP Monitoring (Network Device Polling)
# -----------------------------------------
pysnmp==4.4.12
pyasn1==0.5.1
pysnmp-mibs==0.1.6  # MIB support

# SSH Support (Multi-vendor Device Access)
# -----------------------------------------
paramiko==3.4.0
asyncssh==2.14.1
netmiko==4.2.0  # Multi-vendor SSH library

# Redis (Caching & Session Management)
# -------------------------------------
redis[hiredis]==4.6.0
aioredis==2.0.1

# Background Tasks & Scheduling
# ------------------------------
celery[redis]==5.3.4
flower==2.0.1  # Celery monitoring
celery-beat==2.5.0  # Scheduled tasks

# WebSocket Support (Real-time Updates)
# --------------------------------------
websockets==12.0
python-socketio==5.10.0

# Monitoring & Observability
# --------------------------
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Email Notifications
# -------------------
aiosmtplib==3.0.1
email-validator==2.1.0

# Data Processing & Time-series
# ------------------------------
numpy==1.24.3
pandas==2.1.4
influxdb-client==1.38.0  # Time-series metrics storage

# Utilities
# ---------
tenacity==8.2.3
orjson==3.9.10
python-dateutil==2.8.2
pytz==2023.3
jinja2==3.1.2

# Testing & Development
# ---------------------
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-benchmark==4.0.0
pytest-mock==3.12.0
pytest-json-report==1.5.0
coverage[toml]==7.3.2
fakeredis==2.20.0  # Redis testing
factory-boy==3.3.0  # Test data factories

# Code Quality & Security Scanning
# ---------------------------------
black==23.11.0
flake8==6.1.0
mypy==1.7.1
bandit==1.7.5
safety==3.0.1
pylint==3.0.3

# API Documentation
# -----------------
pydantic-openapi-schema==1.3.0

# Performance Testing
# -------------------
locust==2.19.1  # Load testing

# Additional Network Discovery
# -----------------------------
python-nmap==0.7.1  # Network scanning
scapy==2.5.0  # Packet manipulation for CDP/LLDP
mac-vendor-lookup==0.1.12  # MAC address vendor identification

# Graph & Topology Visualization
# -------------------------------
networkx==3.2.1  # Network topology
graphviz==0.20.1  # Graph rendering

# Logging & Structured Logging
# -----------------------------
structlog==23.2.0
python-json-logger==2.0.7

# Rate Limiting
# -------------
slowapi==0.1.9

# Data Validation & Serialization
# --------------------------------
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# CLI Support
# -----------
click==8.1.7
rich==13.7.0  # Beautiful terminal output

# Deployment & Container Support
# -------------------------------
docker==7.0.0
kubernetes==28.1.0

# Health Checks
# -------------
py-healthcheck==2.0.2

# Compression
# -----------
python-snappy==0.6.1  # Fast compression for metrics

# Additional Dependencies for Full Functionality
# -----------------------------------------------
pyyaml==6.0.1  # YAML configuration support
toml==0.10.2  # TOML configuration support
python-consul==1.1.0  # Service discovery
APScheduler==3.10.4  # Advanced scheduling
pycryptodome==3.19.0  # Additional crypto operations
zstandard==0.22.0  # Zstandard compression
msgpack==1.0.7  # MessagePack serialization
ujson==5.9.0  # Ultra-fast JSON
wrapt==1.16.0  # Decorators and wrappers
contextvars==2.4  # Context variable backport
backoff==2.2.1  # Retry decorator
cachetools==5.3.2  # Caching utilities

# Windows-specific (if needed)
# -----------------------------
# pywin32==306  # Windows API support (uncomment if on Windows)

# =====================================================================
# Installation Instructions:
# -------------------------
# pip install -r chm_requirements.txt
#
# For development environment:
# pip install -r chm_requirements.txt -r requirements-dev.txt
#
# For production deployment with Docker:
# Use multi-stage builds with this file for complete functionality
# =====================================================================